refactor: split napi pm helpers
