+++ crates/ox_content_napi/src/pm/tests.rs
@@ -1,0 +1,14 @@
+use super::commands::convert_command;
+use super::{transform_pm, PmOptions};
+
+fn all(command: &str) -> (String, String, String, String) {
+    (
+        convert_command(command, "npm"),
+        convert_command(command, "pnpm"),
+        convert_command(command, "yarn"),
+        convert_command(command, "bun"),
+    )
+}
+
+mod commands;
+mod transform;
